codec: check if nil in stream explicitly during map key-value decoding
Map values are not addressable.
Yet we have to decode into an addressable value.
However, if the stream contains a nil representation,
then we cannot set that addressable value to nil.
We will need to know this while iterating the map,
and then set the mapping directly to the zero value.

This is implemented here.

Fixes #329