codec: fast atomic updates on a mutating slice cause data race if underlying array is modified
This is why previous code was appending one element at a time in a newly created array.

Revert to same, as it worked pre-v1.1.11.

This is ok, as we are building a cache that is built once and used many times i.e. read-mostly.

Unfortunately, go test -race didn't catch this.

Fixes #336