codec: use tree structure to manage structFieldInfo's.
In the common case where a field is not embedded (anonymous), then
there will be no indirection or allocation in retrieving or tracking the field value.

In the next common case where the type of an anonymous field doesn't have any embedded fields,
there a single pointer indirection is used. Furthermore, we do not need to enlarge the size of
the structFieldInfo by storing a slice in it.