codec: bigendian support for handling numbers/lengths should use byte arrays, not slices
bigenHelper handles ByteOrder operations directly using
arrays of bytes (not slice of bytes).

Since byteorder operations are very common for encoding and decoding
numbers, lengths, etc - it is imperative that this operation is as
fast as possible. Removing indirection (pointer chasing) to look
at up to 8 bytes helps a lot here.

For times where it is expedient to use a slice, delegate to the
bigenstd (equal to the binary.BigEndian value).